Refactored some class locations to make them easier to locate. Refined junit test to get it completely working.
