--- core/src/test/java/gov/nist/secauto/metaschema/core/metapath/cst/MockedDocumentGenerator.java
@@ -1,75 +1,0 @@
-/*
- * SPDX-FileCopyrightText: none
- * SPDX-License-Identifier: CC0-1.0
- */
-
-package gov.nist.secauto.metaschema.core.metapath.cst;
-
-import gov.nist.secauto.metaschema.core.mdm.IDMAssemblyNodeItem;
-import gov.nist.secauto.metaschema.core.mdm.IDMDocumentNodeItem;
-import gov.nist.secauto.metaschema.core.mdm.IDMFieldNodeItem;
-import gov.nist.secauto.metaschema.core.mdm.IDMFlagNodeItem;
-import gov.nist.secauto.metaschema.core.mdm.IDMRootAssemblyNodeItem;
-import gov.nist.secauto.metaschema.core.metapath.item.atomic.IStringItem;
-import gov.nist.secauto.metaschema.core.metapath.item.node.IFlagNodeItem;
-import gov.nist.secauto.metaschema.core.model.IAssemblyDefinition;
-import gov.nist.secauto.metaschema.core.model.IAssemblyInstance;
-import gov.nist.secauto.metaschema.core.model.IFieldInstance;
-import gov.nist.secauto.metaschema.core.model.IFlagDefinition;
-import gov.nist.secauto.metaschema.core.model.IFlagInstance;
-import gov.nist.secauto.metaschema.core.model.ISource;
-import gov.nist.secauto.metaschema.core.qname.IEnhancedQName;
-import gov.nist.secauto.metaschema.core.testing.model.MockedModelTestSupport;
-import gov.nist.secauto.metaschema.core.util.ObjectUtils;
-
-import java.net.URI;
-
-import edu.umd.cs.findbugs.annotations.NonNull;
-
-/**
- * Supports generating a simple Metapath node item-based document model for use
- * in unit tests.
- */
-public class MockedDocumentGenerator {
-  /**
-   * The document base URI used by documents generated by the
-   * {@link #generateDocumentNodeItem()} method.
-   */
-  @NonNull
-  public static final URI BASE_URI = ObjectUtils.notNull(URI.create("https://example.com/resource"));
-  /**
-   * The namespace used by Metaschema module definitions generated by the
-   * {@link #generateDocumentNodeItem()} method.
-   */
-  @NonNull
-  public static final String NS = "http://example.com/ns";
-  /**
-   * The qualified name for the flag node ({@code /Q{http://example.com/ns}flag})
-   * of a document generated by the {@link #generateDocumentNodeItem()} method.
-   */
-  @NonNull
-  public static final IEnhancedQName ROOT_FLAG_QNAME = IEnhancedQName.of("flag");
-
-  /**
-   * Generate a simple Metaschema-based document model.
-   *
-   * @return the document node item for this model
-   */
-  public static IDMFlagNodeItem generateDocumentNodeItem() {
-    MockedModelTestSupport mocking = new MockedModelTestSupport();
-
-    ISource moduleSource = ISource.externalSource("https://example.com/module");
-
-    IFlagDefinition flagDefinition = mocking.flag()
-        .qname(ROOT_FLAG_QNAME)
-        .source(moduleSource)
-        .toDefinition();
-
-    return IDMFlagNodeItem.newInstance(flagDefinition, IStringItem.valueOf("flag"), moduleSource.getStaticContext());
-  }
-
-  private MockedDocumentGenerator() {
-    // disable construction
-  }
-
-}
